requirements.txt
pinecone
requests
datasets
pandas
pyarrow
python-dotenv
sentence-transformers
transformers
pinecone-client